REFACTOR each process has a dedicated module
Why: The project gets to the second step where each process starts to
build methods for each type of data. It is easier to manage each process
with a dedicated module rather than gathering all into a single one.

How: The classes in the clss.py module are moved into the corresponding
modules.

Tags: