REFACTOR Data Validation process
Why: The program was built on the claim view in the first stage. However
the pharmacy view will be introduced; thus, the code was refactored
in the way that it will effectively accomodate both views

How: A new class ViewDataValidation is introduced. This class is the
parent class for both claim and pharmacy view data validation classes.
Also, each process to validate data is constructed as a separate class
to ease the modificaion process, if new ideas are introduced for each
process.

Tags: